Updated DebugControls.kt to accept a viewModel argument so that the buttons can actually modify the game state (Probably need to change this to not just give *everything* to the debug controls this way).
Added extra protections to incrementLevel() after allowing it to accept a levels argument. Protections keep level value within 1 to 1000.

Added debug mutableState Boolean to PlayerViewModel.kt to allow us to toggle visibility of the DebugControls component, as well as a toggleDebug() function.

Added a debugLvlUp() function to PlayerViewModel.kt to use with various DebugControls buttons.

Updated HomeScreen.kt to use debug state from PlayerViewModel.kt to control visibility of a DebugControls() component, and the Settings button to call toggleDebug() when clicked.

Encountered an issue with making edits to the currentWonder in the PlayerViewModel that just have the currentWonder call void functions to modify itself. Currently got around it by making debug functions that only modify the currentWonder also modify the values of money and mana by specific amounts. Will need to address this later.