Added new NavigationButton component.
Updated HomeScreen to use NavigationButton components for the set of buttons at the bottom of the screen.

Updated Settings button, money symbol, and mana symbol to use new assets and made necessary styling fixes.

Restyled the stats UI to wrap money, mana, and the settings button into a Box component to re-arrange the UI to more closely match the original app.